Read timezone from a config file.
Date65 doesn't only display the time, but also sets it "in the system", so it may (in certain scenarios) be of actual use, insteas of just being a demo. But if it's more than a demo, it needs to allow the user to set his timezone.

Now Date65 reads a file named 'date65.tz'. It only reads the first three characters, the first character has to be either '+' or '-'. The second and third character denote the hour deviation from UTC.

The default date65.tz contains '-06' thus matching e.g. Central Standard Time (CST).