Added option to parse URL without resolving the hostname.
In order to make this useful, it is of course necessary to allow access to the parsed hostname. Unfortunately, the only way to do so without much overhead, is to access the name in the generated HTTP header (aka selector). But there, the hostname is terminated with "\r\n\r\n".